// src/HOPPER.Frontend/src/app/api/model/modrinthInstallResultDto.ts
/**
 * HOPPER.API | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ModrinthFailedDto } from './modrinthFailedDto';
import { ModDto } from './modDto';
import { ModrinthAdoptedDto } from './modrinthAdoptedDto';
import { ModrinthSkippedDto } from './modrinthSkippedDto';


export interface ModrinthInstallResultDto { 
    installed: Array<ModDto>;
    adopted: Array<ModrinthAdoptedDto>;
    replaced: Array<ModDto>;
    skipped: Array<ModrinthSkippedDto>;
    failed: Array<ModrinthFailedDto>;
}